fixed delete unit tests, added delete request group tests

+++ APIGiraffe.ApplicationServices/Requests/Commands/DeleteRequestGroup/DeleteRequestGroupCommand.cs
@@ -1,4 +1,5 @@
-﻿using APIGiraffe.Data.UnitOfWork;
+﻿using System;
+using APIGiraffe.Data.UnitOfWork;
 using System.Linq;
 
 namespace APIGiraffe.ApplicationServices.Requests.Commands.DeleteRequestGroup
@@ -15,6 +16,12 @@ namespace APIGiraffe.ApplicationServices.Requests.Commands.DeleteRequestGroup
         public void Execute(int requestGroupId)
         {
             var group = _uow.RequestGroups.Find(requestGroupId);
+
+            if (group == null)
+            {
+                throw new ArgumentOutOfRangeException(nameof(requestGroupId), "Cannot find requestGroup in Db");
+            }
+
             var requests = _uow.Requests.Where(c => c.GroupId == group.Id);
 
             _uow.Requests.RemoveRange(requests);
